working on the authentification and addpost

+++ client/src/services/UserService.js
@@ -1,4 +1,5 @@
 import Api from '@/services/Api'
+import router from '@/router/'
 
 export default {
     register(credentials) {
@@ -11,13 +12,20 @@ export default {
         return Api().get('authentification')
     },
     async checkAuthentification() {
-        try {
-            const response = await this.authentification()
-            return response.data.message
+        if(!localStorage.token) {
+            alert('로그인이 필요합니다. 로그인 창으로 이동합니다.')
+            router.push('login')
         }
-        catch(error) {
-            localStorage.clear()
-            return error.response.data.error
-        }  
+        else {
+            try {
+                const response = await this.authentification()
+                return response.data.user
+            }
+            catch(error) {
+                localStorage.clear()
+                alert('인증과정 중에 문제가 발생했습니다. 로그인 창으로 이동합니다')
+                router.push('login')
+            }  
+        }
     },
 }
